Рефакторинг 3 Убрал Варнинги

+++ Level.cpp
@@ -8,7 +8,7 @@ int Object::GetPropertyInt(std::string name)
 
 float Object::GetPropertyFloat(std::string name)
 {
-	return strtod(properties[name].c_str(), NULL);
+	return float(strtod(properties[name].c_str(), NULL));
 }
 
 std::string Object::GetPropertyString(std::string name)
@@ -92,8 +92,8 @@ bool Level::LoadFromFile(std::string filename)
 		// Если присутствует opacity, то задаем прозрачность слоя, иначе он полностью непрозрачен
 		if (layerElement->Attribute("opacity") != NULL)
 		{
-			float opacity = strtod(layerElement->Attribute("opacity"), NULL);
-			layer.opacity = 255 * opacity;
+			float opacity = float(strtod(layerElement->Attribute("opacity"), NULL));
+			layer.opacity = int(255 * opacity);
 		}
 		else
 		{
@@ -119,8 +119,8 @@ bool Level::LoadFromFile(std::string filename)
 			return false;
 		}
 
-		int x = 0;
-		int y = 0;
+		float x = 0;
+		float y = 0;
 
 		while (tileElement)
 		{
@@ -133,7 +133,7 @@ bool Level::LoadFromFile(std::string filename)
 				sf::Sprite sprite;
 				sprite.setTexture(tilesetImage);
 				sprite.setTextureRect(subRects[subRectToUse]);
-				sprite.setPosition(x * tileWidth, y * tileHeight);
+				sprite.setPosition(float(x * tileWidth), float(y * tileHeight));
 				sprite.setColor(sf::Color(255, 255, 255, layer.opacity));
 
 				layer.tiles.push_back(sprite);
@@ -182,8 +182,8 @@ bool Level::LoadFromFile(std::string filename)
 				{
 					objectName = objectElement->Attribute("name");
 				}
-				int x = atoi(objectElement->Attribute("x"));
-				int y = atoi(objectElement->Attribute("y"));
+				float x = float(atoi(objectElement->Attribute("x")));
+				float y = float(atoi(objectElement->Attribute("y")));
 
 				int width, height;
 
@@ -211,10 +211,10 @@ bool Level::LoadFromFile(std::string filename)
 				object.sprite = sprite;
 
 				sf::Rect <float> objectRect;
-				objectRect.top = y;
-				objectRect.left = x;
-				objectRect.height = height;
-				objectRect.width = width;
+				objectRect.top = float(y);
+				objectRect.left = float(x);
+				objectRect.height = float(height);
+				objectRect.width = float(width);
 				object.rect = objectRect;
 
 				// "Переменные" объекта
@@ -257,7 +257,7 @@ bool Level::LoadFromFile(std::string filename)
 Object Level::GetObject(std::string name)
 {
 	// Только первый объект с заданным именем
-	for (int i = 0; i < objects.size(); i++)
+	for (size_t i = 0; i < objects.size(); i++)
 		if (objects[i].name == name)
 			return objects[i];
 }
@@ -266,7 +266,7 @@ std::vector<Object> Level::GetObjects(std::string name)
 {
 	// Все объекты с заданным именем
 	std::vector<Object> vec;
-	for (int i = 0; i < objects.size(); i++)
+	for (size_t i = 0; i < objects.size(); i++)
 		if (objects[i].name == name)
 			vec.push_back(objects[i]);
 
@@ -288,7 +288,7 @@ sf::Vector2i Level::GetTileSize()
 void Level::Draw(sf::RenderWindow &window)
 {
 	// Рисуем все тайлы (объекты НЕ рисуем!)
-	for (int layer = 0; layer < layers.size(); layer++)
-		for (int tile = 0; tile < layers[layer].tiles.size(); tile++)
+	for (size_t layer = 0; layer < layers.size(); layer++)
+		for (size_t tile = 0; tile < layers[layer].tiles.size(); tile++)
 			window.draw(layers[layer].tiles[tile]);
 }
